adding error handling to shell script

+++ src/commons/scripts/create-cypress-env.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ -z "$BS_EMAIL" ] || [ -z "$BS_PASSWORD" ] || [ -z "$EXTERNAL_IP" ]; then
+  echo "Error: Environment variables BS_EMAIL, BS_PASSWORD and EXTERNAL_IP must be defined."
+  exit 1
+fi
+
 rm -f cypress.env.json
 
 env_vars='{
@@ -8,5 +13,10 @@ env_vars='{
   "EXTERNAL_IP": "'"${EXTERNAL_IP}"'"
 }'
 
-echo "$env_vars" > cypress.env.json
+if ! echo "$env_vars" > cypress.env.json; then
+  echo "Error: fail in cypress.env.json file creation."
+  exit 1
+fi
+
 echo "File cypress.env.json generated with success."
+exit 0
